requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor==6.7.2
django-filter==25.1
django-js-asset==3.0.1
djangorestframework==3.15.2
Markdown==3.7
pillow==11.1.0
psycopg2==2.9.10
python-dotenv==1.0.1
router==0.1
sqlparse==0.5.3
tzdata==2025.1
